Top-level functions

Main functionality for TMI.

TMI.trackpathways — Function

function trackpathways(TMIversion,latbox,lonbox)
Track the pathways of a user-defined water mass.
 Steps: (a) define the water mass by a rectangular surface patch dyed with passive tracer concentration of         (b) propagate the dye with the matrix A, with the result being the fraction of water originating from the surface region.
 See Section 2b of Gebbie & Huybers 2010, esp. eqs. (15)-(17).

Arguments

TMIversion: version of TMI water-mass/circulation model

latbox: min and max latitude of box

lonbox: min and max longitude of box

γ: TMI grid

Output

c: fraction of water from surface source

TMI.watermassdistribution — Function

function watermassdistribution(TMIversion,latbox,lonbox)
Track the pathways of a user-defined water mass.
 Steps: (a) define the water mass by an oceanographically-relevant surface patch dyed with passive tracer concentration of one
     (b) propagate the dye with the matrix A, with the result being the fraction of water originating from the surface region.
 See Section 2b of Gebbie & Huybers 2010, esp. eqs. (15)-(17).

Arguments

TMIversion: version of TMI water-mass/circulation model

Alu: LU decomposition of water-mass matrix A

region: name of pre-defined surface region

γ: TMI grid

Output

g: water-mass fraction

TMI.volumefilled — Function

function volumefilled(TMIversion)
Find the ocean volume that has originated from each surface box.
 This is equivalent to solving a sensitivity problem:
 The total volume is V = vᵀ c , where v is the volume of each box 
 and c is the fraction of volume from a given source which
 satisfies the equation A c = d.                     
 Next, dV/d(d) = A⁻ᵀ v, and dV/d(d) is exactly the volume originating from each source.

 See Section 3 and Supplementary Section 4, Gebbie & Huybers 2011.

Arguments

TMIversion: version of TMI water-mass/circulation model

Alu: LU decomposition of water-mass matrix A

γ: TMI.grid

Output

volume: log10 of global ocean volume filled by a surface region, exists at surface, therefore given BoundaryCondition type

TMI.surfaceorigin — Function

function surfaceorigin(TMIversion,loc)
 Find the surface origin of water for some interior box 
 This is equivalent to solving a sensitivity problem:
 The mass fraction at a location `loc` of interest is 
`c[loc] = δᵀ c`, where `δ` samples the location of the global mass-fraction variable, c.
Then the sensitivity of `c[loc]` is: d(c[loc])/d(d) = A⁻ᵀ δ.
The derivative is solved using the constraint: Ac = d.
The sensitivity is exactly the mass fraction originating from each source.      
This problem is mathematically similar to determining how the ocean is filled.

Arguments

loc: location (lon,lat,depth) of location of interest

Alu: LU decomposition of water-mass matrix A

γ: TMI grid

Output

origin: surface map of fraction of source water for a given location, log10 of effective depth, in terms of a BoundaryCondition

TMI.meanage — Function

function meanage(TMIversion,Alu,γ)
Mean or ideal age

Arguments

TMIversion: version of TMI water-mass/circulation model

Alu: LU decomposition of water-mass matrix A

γ: TMI grid

Output

a: mean age [yr]