TMI

Documentation for TMI.

TMI.cartesianindexXYZ

TMI.cartesianindexZYX

TMI.config

TMI.depthindex

TMI.download

TMI.dyeplot

TMI.fld2vec

TMI.gdriveurl

TMI.gridprops

TMI.horizontaldistance

TMI.linearindexXYZ

TMI.misfit_gridded_data

TMI.nearestneighbor

TMI.nearestneighbormask

TMI.plotextent

TMI.readtracer

TMI.regeneratedphosphate

TMI.section

TMI.surfaceindex

TMI.surfaceorigin

TMI.surfacepatch

TMI.tracerinit

TMI.trackpathways

TMI.updatelinearindex

TMI.vec2fld

TMI.volumefilled

TMI.watermassmatrix

TMI.watermassmatrixXYZ

TMI.watermassmatrixZYX

TMI.Γ

TMI.cartesianindexXYZ — Method

function cartesianindexXYZ(wet)
Read and assemble the grid coordinates
according to a 3D tracer in x,y,z order

Arguments

wet: BitArray logical mask for wet points

Output

I: 3D Cartesian indices

TMI.cartesianindexZYX — Method

function cartesianindexZYX(file)
Read and assemble the grid coordinates
according to the legacy MATLAB code (z,y,x order).

Arguments

file: TMI NetCDF file name

Output

grid: TMI grid coordinates

TMI.config — Method

function config(url,inputdir)

Arguments

TMIversion: TMI version for water-mass/circulation model

Output

A: TMI steady-state water-mass matrix

Alu: LU decomposition of A

γ: TMI grid properties

TMIfile: TMI file name

TMI.depthindex — Method

function depthindex(I) 

Get the k-index (depth level) from the Cartesian index

TMI.download — Method

function download(url,inputdir)
Read and assemble all TMI inputs.

Arguments

url: Google Drive location of TMI input

inputdir: input directory location to store file

Output

none

TMI.dyeplot — Method

function dyeplot
Plot of dye in ocean

Arguments

lat: latitude arrays

depth: depth array

vals: lat x depth value array

lims: contour levels

TMI.fld2vec — Method

function fld2vec
Transfer 3D field with accounting for ocean bathymetry to a vector without land points

Arguments

field: field in 3d form including land points (NaN)

I: cartesian indices of ocean points

Output

vector: field in vector form (no land points)

TMI.gdriveurl — Method

function gdriveurl(TMIversion)
placeholder function to give location (URL) of Google Drive input
in the future, consider a struct or Dict that describes all TMI versions.

Arguments

TMIversion: version of TMI water-mass/circulation model

Output

url: location (URL) for download

TMI.gridprops — Method

function gridprops(file)
Read and assemble the grid properties.

Arguments

file: TMI NetCDF file name

Output

grid: TMI grid coordinates

TMI.horizontaldistance — Method

function horizontaldistance(loc,γ)
return the Cartesian index and linear index 
of the nearest N neighbors

Arguments

loc: 3-tuple of lon,lat,depth location

γ: TMI.grid

Output

hordist: horizontal distance to nearest tracer grid points

TMI.linearindexXYZ — Method

function linearindexXYZ(file)
Read and assemble the grid coordinates.

Arguments

wet: 3D mask for wet points

Output

R: array of linear indices, but not a LinearIndices type

TMI.misfit_gridded_data — Method

function misfit_gridded_data(u,Alu,y,d,Wⁱ,wet)
squared model-data misfit
controls are a vector input for Optim.jl

Arguments

u: controls, vector format

Alu: LU decomposition of water-mass matrix

y: observations on grid

d: model constraints

Wⁱ: inverse of W weighting matrix for observations

wet: BitArray ocean mask

Output

J: cost function of sum of squared misfits

gJ: derivative of cost function wrt to controls

TMI.nearestneighbor — Method

function nearestneighbor(loc,γ)
return the Cartesian index and linear index 
of the nearest N neighbors

Arguments

loc: 3-tuple of lon,lat,depth location

γ: TMI.grid

Output

Inn: Cartesian indices of nearest neighbor

Rnn: linear indices of nearest neighbor

TMI.nearestneighbormask — Method

function nearestneighbormask
Make a 3D tracer field that is 1 at location 
of nearest neighbor, 0 elsewhere

Arguments

loc: location in a 3-tuple (lon,lat,depth)

γ: TMI.grid

Output

δ: nearest neighbor mask 3D field

TMI.plotextent — Method

function plotextent
Generate image showing user-specified ROI

Arguments

latbox: in format [latstart, latstop]

lonbox: in format [lonstart, lonstop]

TMI.readtracer — Method

function readtracer(file,tracername)
Read and assemble the water-mass matrix.

Arguments

file: TMI NetCDF file name

tracername: name of tracer

Output

c: 3D tracer field

TMI.regeneratedphosphate — Method

function regeneratedphosphate(TMIversion)
Regenerated (i.e., accumulated, remineralized) phosphate

Arguments

TMIversion: version of TMI water-mass/circulation model

Output

PO₄ᴿ: regenerated phosphate

γ: TMI grid

TMI.section — Method

function section
View latitude-depth slice of field

Arguments

c: 3d tracer field

lon: longitude of section

γ: TMI.grid

Output

csection: 2d slice of field

TMI.surfaceindex — Method

function surfaceindex(I) 

Get the vector-index where depth level == 1 and it is ocean.

TMI.surfaceorigin — Method

function surfaceorigin(TMIversion,loc)

Arguments

TMIversion: version of TMI water-mass/circulation model

loc: location (lon,lat,depth) of location of interest

Output

origin: surface map of fraction of source water for a given location

γ: TMI grid

TMI.surfacepatch — Method

function surfacepatch
Make a surface boundary condition
with a rectangular patch

Arguments

lonbox: longitudes of box edges

latbox: latitudes of box edges

γ: TMI.grid

Output

d: vector that describes surface patch

TMI.tracerinit — Function

function tracerinit(wet,ltype=Float64)
  initialize tracer field on TMI grid
perhaps better to have a tracer struct and constructor

Arguments

wet::BitArray mask of ocean points

Output

d:: 3d tracer field with NaN on dry points

TMI.trackpathways — Method

function trackpathways(TMIversion,latbox,lonbox)
Fraction of water originating from surface box

Arguments

TMIversion: version of TMI water-mass/circulation model

latbox: min and max latitude of box

lonbox: min and max longitude of box

Output

c: fraction of water from surface source

γ: TMI grid

TMI.updatelinearindex — Method

function updatelinearindex(izyx,Izyx,R)
Linear index translated from z,y,x to x,y,z accounting

Arguments

izyx: index of interest in z,y,x accounting

Izyx: wet Cartesian Index for z,y,x

R: Linear indices for x,y,z

Output

ixyz: index of interest in x,y,z accounting

TMI.vec2fld — Method

function vec2fld
Transfer a vector to a 3D field with accounting for ocean bathymetry

Arguments

vector: field in vector form (no land points)

I: cartesian indices of ocean points

Output

field: field in 3d form including land points (NaN)

TMI.volumefilled — Method

function volumefilled(TMIversion)

Arguments

TMIversion: version of TMI water-mass/circulation model

Output

volume: global ocean volume filled by a surface region

TMI.watermassmatrix — Method

function watermassmatrix(file)
Read and assemble the water-mass matrix.

Arguments

file: TMI NetCDF file name

Output

A: water-mass matrix

TMI.watermassmatrixXYZ — Method

    function watermassmatrixXYZ(file,R)
Read and assemble the water-mass matrix from MATLAB.
Transfer to updated x,y,z version

Arguments

file: TMI NetCDF file name

γ: TMI grid

Output

A: water-mass matrix

TMI.watermassmatrixZYX — Method

function watermassmatrixZYX(file)
Read and assemble the water-mass matrix.
Legacy version from MATLAB.

Arguments

file: TMI NetCDF file name

Output

A: water-mass matrix

TMI.Γ — Method

function Γ(tracer2D,γ)
turn 2D surface field into 3D field with zeroes below surface

Arguments

tracer2D:: 2D surface tracer field

wet::BitArray mask of ocean points

Output

tracer3D:: 3d tracer field with NaN on dry points