TMI

Documentation for TMI.

TMI.boundarymatrix

TMI.boundarymatrix2nc

TMI.cartesianindex

TMI.cartesianindex

TMI.cellarea

TMI.cellvolume

TMI.ces_ncwrite

TMI.circulationmatrix

TMI.circulationmatrix

TMI.circulationmatrix2nc

TMI.config2nc

TMI.config_from_mat

TMI.config_from_nc

TMI.control2state

TMI.control2state

TMI.control2state!

TMI.control2state!

TMI.costfunction

TMI.costfunction!

TMI.costfunction_obs

TMI.costfunction_obs

TMI.costfunction_obs!

TMI.costfunction_obs!

TMI.depthindex

TMI.dyeplot

TMI.fieldsatts

TMI.fld2vec

TMI.grid2nc

TMI.griddicts

TMI.gridprops

TMI.horizontaldistance

TMI.interpindex

TMI.interpweights

TMI.latindex

TMI.linearindex

TMI.lonindex

TMI.matfields2nc

TMI.matrix_zyx2xyz

TMI.maturl

TMI.ncurl

TMI.nearestneighbor

TMI.nearestneighbormask

TMI.observe

TMI.optim2nc

TMI.oxygen

TMI.plotextent

TMI.readtracer

TMI.regeneratedphosphate

TMI.regions2nc

TMI.section

TMI.steady_inversion

TMI.steady_inversion

TMI.steady_inversion

TMI.steadyclimatology

TMI.surface_oxygensaturation

TMI.surfaceindex

TMI.surfaceorigin

TMI.surfacepatch

TMI.surfaceregion

TMI.synthetic_observations

TMI.synthetic_observations

TMI.tracerinit

TMI.tracerinit

TMI.trackpathways

TMI.updatelinearindex

TMI.varying!

TMI.vec2fld

TMI.volumefilled

TMI.watermassdistribution

TMI.watermassmatrix

TMI.wetlocation

TMI.boundarymatrix — Method

    function boundarymatrix(file,γ)
Read and assemble the boundary matrix from MATLAB.
Transfer to updated x,y,z version

Arguments

file: TMI MATLAB file name

γ: TMI grid

Output

B: boundary condition matrix

TMI.cartesianindex — Method

function cartesianindex(wet)
Read and assemble the grid coordinates
according to a 3D tracer in x,y,z order

Arguments

wet: BitArray logical mask for wet points

Output

I: 3D Cartesian indices

TMI.cartesianindex — Method

function cartesianindex(file)
Read and assemble the grid coordinates
according to the legacy MATLAB code (z,y,x order).

Arguments

file: TMI NetCDF file name

Output

I: TMI Cartesian index for wet points

TMI.cellarea — Method

Horizontal area of grid cell

TMI.cellvolume — Method

Volume of each grid cell.

TMI.ces_ncwrite — Method

function ces_ncwrite(γ,time,sol_array)
Write .nc file output for commonerasim.jl

Arguments

γ:

time: vector of time values

sol_array: solution array in form time x lat x lon x depth - must match γ + time

Output

saves .nc file titled "ces_output.nc" in data array

TMI.circulationmatrix — Method

function circulationmatrix(file,A,γ)
Read and assemble the circulation matrix from the efficient storage of A and F₀ variables.

Arguments

file: TMI MATLAB file name

A: TMI water-mass matrix

γ: TMI grid

Output

L: circulation matrix in xyz format

TMI.circulationmatrix — Method

function circulationmatrix(file,γ)
Read and assemble the circulation matrix from MATLAB.
Transfer to updated x,y,z version

Arguments

file: TMI MATLAB file name

γ: TMI grid

Output

L: circulation matrix in xyz format

TMI.config2nc — Method

Save TMI configuration to NetCDF format for non-proprietary access

TMI.config_from_mat — Method

Configure TMI environment from original MATLAB output

TMI.config_from_nc — Method

function config_from_nc(TMIversion)
Configure TMI environment from NetCDF input file.

Arguments

TMIversion: TMI version for water-mass/circulation model

Output

A: TMI steady-state water-mass matrix

Alu: LU decomposition of A

γ: TMI grid properties

TMIfile: TMI file name

TMI.control2state! — Method

function control2state!(c,u,γ)
Add surface control vector to existing 3D field

Arguments

c:: state field, 3d tracer field with NaN on dry points, modified by function

u:: surface control vector

wet::BitArray mask of ocean points

TMI.control2state! — Method

function control2state!(c,u,γ)
Add surface control vector to existing 3D field

Arguments

c:: state field, 3d tracer field with NaN on dry points, modified by function

u:: surface control vector

wet::BitArray mask of ocean points

TMI.control2state — Method

function control2state(tracer2D,γ)
turn 2D surface field into 3D field with zeroes below surface

Arguments

tracer2D:: 2D surface tracer field

wet::BitArray mask of ocean points

Output

tracer3D:: 3d tracer field with NaN on dry points

TMI.control2state — Method

function control2state(u,γ)
turn surface control vector into 3D field with zeroes below surface

Arguments

u:: surface control vector

wet::BitArray mask of ocean points

Output

tracer3D:: 3d tracer field with NaN on dry points

TMI.costfunction! — Method

function costfunction!(J,gJ,u,Alu,dfld,yfld,Wⁱ,wis,Q⁻,γ)
squared model-data misfit for pointwise data
controls are a vector input for Optim.jl
Issue: couldn't figure out how to nest with costfunction_obs!
Issue: why are wis and locs both needed?

Arguments

J: cost function of sum of squared misfits

gJ: derivative of cost function wrt to controls

u: controls, vector format

Alu: LU decomposition of water-mass matrix

dfld: model constraints

y: pointwise observations

Wⁱ: inverse of W weighting matrix for observations

wis: weights for interpolation (data sampling, E)

locs: data locations (lon,lat,depth)

Q⁻: weights for control vector

γ: grid

TMI.costfunction — Method

function costfunction(J,gJ,u,Alu,dfld,yfld,Wⁱ,wis,Q⁻,γ)
squared model-data misfit for pointwise data
controls are a vector input for Optim.jl
Issue: couldn't figure out how to nest with costfunction_obs!

Arguments

u: controls, vector format

Alu: LU decomposition of water-mass matrix

dfld: model constraints

y: pointwise observations

Wⁱ: inverse of W weighting matrix for observations

wis: weights for interpolation (data sampling, E)

locs: data locations (lon,lat,depth)

Q⁻: weights for control vector

γ: grid

Output

J: cost function of sum of squared misfits

gJ: derivative of cost function wrt to controls

TMI.costfunction_obs! — Method

function costfunction_obs(u,Alu,y,d,Wⁱ,wet)
squared model-data misfit for gridded data
controls are a vector input for Optim.jl

Arguments

J: cost function of sum of squared misfits

gJ: derivative of cost function wrt to controls

u: controls, vector format

Alu: LU decomposition of water-mass matrix

d: model constraints

y: observations on grid

Wⁱ: inverse of W weighting matrix for observations

γ: grid

TMI.costfunction_obs! — Method

function costfunction_obs!(J,gJ,u,Alu,dfld,yfld,Wⁱ,wis,locs,γ)
squared model-data misfit for pointwise data
controls are a vector input for Optim.jl

Arguments

J: cost function of sum of squared misfits

gJ: derivative of cost function wrt to controls

u: controls, vector format

Alu: LU decomposition of water-mass matrix

dfld: model constraints

y: pointwise observations

Wⁱ: inverse of W weighting matrix for observations

wis: weights for interpolation (data sampling, E)

locs: data locations (lon,lat,depth)

γ: grid

TMI.costfunction_obs — Method

function costfunction_obs(u,Alu,dfld,yfld,Wⁱ,γ)
squared model-data misfit for gridded data
controls are a vector input for Optim.jl

Arguments

u: controls, vector format

Alu: LU decomposition of water-mass matrix

y: observations on grid

d: model constraints

Wⁱ: inverse of W weighting matrix for observations

wet: BitArray ocean mask

Output

J: cost function of sum of squared misfits

gJ: derivative of cost function wrt to controls

TMI.costfunction_obs — Method

function costfunction_obs(u,Alu,dfld,yfld,Wⁱ,wis,locs,γ)
squared model-data misfit for pointwise data
controls are a vector input for Optim.jl

Arguments

u: controls, vector format

Alu: LU decomposition of water-mass matrix

y: pointwise observations

d: model constraints

Wⁱ: inverse of W weighting matrix for observations

wis: weights for interpolation

locs: data locations

γ: grid

Output

J: cost function of sum of squared misfits

gJ: derivative of cost function wrt to controls

TMI.depthindex — Method

function depthindex(I) 
Get the k-index (depth level) from the Cartesian index

TMI.dyeplot — Function

function dyeplot
Plot of dye in ocean

Arguments

lat: latitude arrays

depth: depth array

vals: lat x depth value array

lims: contour levels

TMI.fld2vec — Method

function fld2vec
Transfer 3D field with accounting for ocean bathymetry to a vector without land points.
This is done more easily with a BitArray mask, i.e., vector = field[mask].
This function may be removed in the future.

Arguments

field: field in 3d form including land points (NaN)

I: cartesian indices of ocean points

Output

vector: field in vector form (no land points)

TMI.gridprops — Method

function gridprops(file)
Read and assemble the grid properties.

Arguments

file: TMI NetCDF file name

Output

grid: TMI grid coordinates

TMI.horizontaldistance — Method

function horizontaldistance(loc,γ)
return the Cartesian index and linear index 
of the nearest N neighbors

Arguments

loc: 3-tuple of lon,lat,depth location

γ: TMI.grid

Output

hordist: horizontal distance to nearest tracer grid points

TMI.interpindex — Method

function interpindex(loc,γ) Weights for linear interpolation. The derivative of linear interpolation is needed in sensitivity studies. ReverseDiff.jl could find this quantity automatically. Instead we dig into the Interpolations.jl package to find the weights that are effectively the partial derivatives of the function.

Arguments

c: a temporary tracer field, would be nice to make it unnecessary

loc: (lon,lat,depth) tuple of a location of interest

γ: TMI grid

Output

δ: weights on a 3D tracer field grid

TMI.interpweights — Method

function interpweights(loc,γ) Weights for linear interpolation. The derivative of linear interpolation is needed in sensitivity studies. ReverseDiff.jl could find this quantity automatically. Instead we dig into the Interpolations.jl package to find the weights that are effectively the partial derivatives of the function.

Arguments

loc: (lon,lat,depth) tuple of a location of interest

γ: TMI grid

Output

δ: weights on a 3D tracer field grid

TMI.latindex — Method

function latindex(I) 
Get the j-index (latitude index) from the Cartesian index

TMI.linearindex — Method

function linearindex(wet)
Read and assemble the grid coordinates.

Arguments

wet: 3D mask for wet points

Output

R: array of linear indices, but not a LinearIndices type

TMI.lonindex — Method

function lonindex(I) 
Get the i-index (lon index) from the Cartesian index

TMI.matrix_zyx2xyz — Method

    function matrix_zyx2xyz(TMIfile,Azyx,γ)

Transfer zyx format water-mass matrix A to xyz format

Arguments

Azyx: water-mass matrix in zyx format

γ: TMI grid

Output

Axyz: water-mass matrix in xyz format

TMI.nearestneighbor — Function

function nearestneighbor(loc,γ)
return the Cartesian index and linear index 
of the nearest N neighbors

Arguments

loc: 3-tuple of lon,lat,depth location

γ: TMI.grid

Output

Inn: Cartesian indices of nearest neighbor

#- Rnn: linear indices of nearest neighbor, Removed from code

TMI.nearestneighbormask — Function

function nearestneighbormask
Make a 3D tracer field that is 1 at location 
of nearest neighbor, 0 elsewhere

Arguments

loc: location in a 3-tuple (lon,lat,depth)

γ: TMI.grid

Output

δ: nearest neighbor mask 3D field

TMI.observe — Method

function observe
Take a observation at location given by weights wis

TMI.oxygen — Method

Reconstruct dissolved oxygen (that doesn't exist in TMI product) by assuming same oxygen saturation fraction as modern

TMI.plotextent — Method

function plotextent
Generate image showing user-specified ROI

Arguments

latbox: in format [latstart, latstop]

lonbox: in format [lonstart, lonstop]

TMI.readtracer — Method

function readtracer(file,tracername)
Read a tracer field from NetCDF.

Arguments

file: TMI NetCDF file name

tracername: name of tracer

Output

c: 3D tracer field

TMI.regeneratedphosphate — Method

function regeneratedphosphate(TMIversion,Alu,γ)
Regenerated (i.e., accumulated, remineralized) phosphate

Arguments

TMIversion: version of TMI water-mass/circulation model

Alu: LU decomposition of water-mass matrix A

γ: TMI grid

Output

PO₄ᴿ: regenerated phosphate

TMI.section — Method

function section
View latitude-depth slice of field

Arguments

c: 3d tracer field

lon: longitude of section

γ: TMI.grid

Output

csection: 2d slice of field

TMI.steady_inversion — Method

function steady_inversion(Alu,d,γ.wet)
invert for a steady-state tracer distribution

Arguments

Alu: LU decomposition of water-mass matrix

d: model constraints

wet: BitArray ocean mask

Output

c: steady-state tracer distribution

TMI.steady_inversion — Method

function steady_inversion(Alu,dsfc,qint,γ.wet)
invert for a steady-state tracer distribution

Arguments

Alu: LU decomposition of water-mass matrix

dsfc: surface boundary condition

q: interior sources/sinks of phosphate

r: stochiometric ratio of tracer:phosphate

wet: BitArray ocean mask

Output

c: steady-state tracer distribution

TMI.steady_inversion — Method

function steady_inversion(u,Alu,d,γ.wet)
invert for a steady-state tracer distribution

Arguments

u: controls, vector format

Alu: LU decomposition of water-mass matrix

d: model constraints

wet: BitArray ocean mask

Output

c: steady-state tracer distribution

TMI.steadyclimatology — Method

function steadyclimatology(u₀,fg!,iterations) Find the distribution of a tracer given: (a) the pathways described by A or its LU decomposition Alu, (b) first-guess boundary conditions and interior sources given by d₀, (c) perturbations to the surface boundary condition u₀ that best fits observations, y, according to the cost function, J = (ỹ - y)ᵀ W⁻¹ (ỹ - y) subject to Aỹ = d₀ + Γ u₀. W⁻ is a (sparse) weighting matrix. See Supplementary Section 2, Gebbie & Huybers 2011.

Arguments

u₀:

fg!: compute cost function and gradient in place

iterations: number of optimization iterations

TMI.surface_oxygensaturation — Method

Surface oxygen saturation value and fraction of saturation value in field

TMI.surfaceindex — Method

function surfaceindex(I) 
Get the vector-index where depth level == 1 and it is ocean.

TMI.surfaceorigin — Method

function surfaceorigin(TMIversion,loc)
 Find the surface origin of water for some interior box 
 This is equivalent to solving a sensitivity problem:
 The mass fraction at a location `loc` of interest is 
`c[loc] = δᵀ c`, where `δ` samples the location of the global mass-fraction variable, c.
Then the sensitivity of `c[loc]` is: d(c[loc])/d(d) = A⁻ᵀ δ.
The derivative is solved using the constraint: Ac = d.
The sensitivity is exactly the mass fraction originating from each source.      
This problem is mathematically similar to determining how the ocean is filled.

Arguments

loc: location (lon,lat,depth) of location of interest

Alu: LU decomposition of water-mass matrix A

γ: TMI grid

Output

origin: surface map of fraction of source water for a given location

TMI.surfacepatch — Method

function surfacepatch
Make a surface boundary condition
with a rectangular patch

Arguments

lonbox: longitudes of box edges

latbox: latitudes of box edges

γ: TMI.grid

Output

d: vector that describes surface patch

TMI.synthetic_observations — Method

function synthetic_observations(TMIversion,variable,locs)
Synthetic observations that are a contaminated version of real observations
This version: observations with random (uniform) spatial sampling

Arguments

TMIversion::String: version of TMI water-mass/circulation model

variable::String: variable name to use as template

N: number of observations

Output

y: contaminated observations on 3D grid

W⁻: appropriate weighting (inverse covariance) matrix for these observations,

ytrue: uncontaminated observations, 3D field

locs: 3-tuples of locations for observations

wis: weighted indices for interpolation to locs sites

TMI.synthetic_observations — Method

function synthetic_observations(TMIversion,variable)
Synthetic observations that are a contaminated version of real observations
This version: gridded observations

Arguments

TMIversion::String: version of TMI water-mass/circulation model

variable::String: variable name to use as template

Output

y: contaminated observations on 3D grid

W⁻: appropriate weighting (inverse covariance) matrix for these observations,

θtrue: real observations, 3D field

TMI.tracerinit — Function

function tracerinit(wet,ltype=Float64)
  initialize tracer field on TMI grid
perhaps better to have a tracer struct and constructor

Arguments

wet::BitArray mask of ocean points

ltype:: optional type argument, default=Float64

Output

d:: 3d tracer field with NaN on dry points

TMI.tracerinit — Method

function tracerinit(wet,vec,I)
      initialize tracer field on TMI grid
    perhaps better to have a tracer struct and constructor

Arguments

wet:: BitArray mask of ocean points

vec:: vector of values at wet points

I:: Cartesian Index for vector

Output

field:: 3d tracer field with NaN on dry points

TMI.trackpathways — Method

function trackpathways(TMIversion,latbox,lonbox)
Track the pathways of a user-defined water mass.
 Steps: (a) define the water mass by a rectangular surface patch dyed with passive tracer concentration of         (b) propagate the dye with the matrix A, with the result being the fraction of water originating from the surface region.
 See Section 2b of Gebbie & Huybers 2010, esp. eqs. (15)-(17).

Arguments

TMIversion: version of TMI water-mass/circulation model

latbox: min and max latitude of box

lonbox: min and max longitude of box

γ: TMI grid

Output

c: fraction of water from surface source

TMI.updatelinearindex — Method

function updatelinearindex(izyx,Izyx,R)
Linear index translated from z,y,x to x,y,z accounting

Arguments

izyx: index of interest in z,y,x accounting

Izyx: wet Cartesian Index for z,y,x

R: Linear indices for x,y,z

Output

ixyz: index of interest in x,y,z accounting

TMI.varying! — Method

function varying!(du, u, p, t)
ODE function for varying boundary cond
Sets up dc/dt = L*C + B*f to be solved

Arguments

du: dc/dt (must have this name for DifferentialEquations.jl to work

u: C, what we are solving for

p: parameters for diffeq - must hold specified vars

t: time we are solving for (automatically determined by DE.jl)

Output

du: numerical value of LC+Bf, vector of size 74064 for 4°

TMI.vec2fld — Method

function vec2fld
Transfer a vector to a 3D field with accounting for ocean bathymetry

Arguments

vector: field in vector form (no land points)

I: cartesian indices of ocean points

Output

field: field in 3d form including land points (NaN)

TMI.volumefilled — Method

function volumefilled(TMIversion)
Find the ocean volume that has originated from each surface box.
 This is equivalent to solving a sensitivity problem:
 The total volume is V = vᵀ c , where v is the volume of each box 
 and c is the fraction of volume from a given source which
 satisfies the equation A c = d.                     
 Next, dV/d(d) = A⁻ᵀ v, and dV/d(d) is exactly the volume originating from each source.

 See Section 3 and Supplementary Section 4, Gebbie & Huybers 2011.

Arguments

TMIversion: version of TMI water-mass/circulation model

Alu: LU decomposition of water-mass matrix A

γ: TMI.grid

Output

volume: global ocean volume filled by a surface region

TMI.watermassdistribution — Method

function watermassdistribution(TMIversion,latbox,lonbox)
Track the pathways of a user-defined water mass.
 Steps: (a) define the water mass by an oceanographically-relevant surface patch dyed with passive tracer concentration of one
     (b) propagate the dye with the matrix A, with the result being the fraction of water originating from the surface region.
 See Section 2b of Gebbie & Huybers 2010, esp. eqs. (15)-(17).

Arguments

TMIversion: version of TMI water-mass/circulation model

Alu: LU decomposition of water-mass matrix A

region: name of pre-defined surface region

γ: TMI grid

Output

g: water-mass fraction

TMI.watermassmatrix — Method

function watermassmatrix(file)
Read and assemble the water-mass matrix.

Arguments

file: TMI NetCDF or MATLAB file name

Output

A: water-mass matrix

TMI.wetlocation — Method

function wetlocation(γ)
Get (lon,lat,depth) tuples of wet locations.
Allow a location to be wet if at least one out of 8 nearby gridpoints is wet.
Certainly "wet" gridpoints could be defined more strictly.

Arguments

γ: TMI.grid

Output

loc: lon,lat,depth

TMI.boundarymatrix2nc — Method

Save boundary matrix for transient model to NetCDF file

TMI.circulationmatrix2nc — Method

Save circulation matrix L to NetCDF file.

TMI.fieldsatts — Method

All variable names and attributes. Useful for writing NetCDF files.

TMI.grid2nc — Method

Put grid properties (Cartesian index) into NetCDF file

TMI.griddicts — Method

Save grid dictionaries of attributes for writing to NetCDF file

TMI.matfields2nc — Method

Read 3D fields from mat file and save to NetCDF file.

TMI.maturl — Method

function maturl(TMIversion)
Find *mat file here.
placeholder function to give location (URL) of Google Drive input
in the future, consider a struct or Dict that describes all TMI versions.

Arguments

TMIversion: version of TMI water-mass/circulation model

Output

url: location (URL) for download

TMI.ncurl — Method

function ncurl(TMIversion)
placeholder function to give location (URL) of NetCDF Google Drive input
in the future, consider a struct or Dict that describes all TMI versions.

Arguments

TMIversion: version of TMI water-mass/circulation model

Output

url: location (URL) for download

TMI.optim2nc — Method

Save optimization parameters to NetCDF file)

Future considerations: split into 2 functions

read from mat

save to nc

TMI.regions2nc — Method

Read vectors from mat file, translate to 3D, and save surface field to NetCDF file.

TMI.surfaceregion — Method

Read an oceanographically-relevant surface region from NetCDF file. (Also could be read from mat file.)