TMI

Documentation for TMI.

TMI.cartesianindexXYZ

TMI.cartesianindexZYX

TMI.config

TMI.control2state

TMI.control2state

TMI.control2state!

TMI.control2state!

TMI.costfunction

TMI.costfunction!

TMI.costfunction_obs

TMI.costfunction_obs

TMI.costfunction_obs!

TMI.costfunction_obs!

TMI.depthindex

TMI.download

TMI.dyeplot

TMI.filterdata

TMI.fld2vec

TMI.gdriveurl

TMI.gridprops

TMI.horizontaldistance

TMI.interpindex

TMI.interpweights

TMI.linearindexXYZ

TMI.nearestneighbor

TMI.nearestneighbormask

TMI.plotextent

TMI.readtracer

TMI.regeneratedphosphate

TMI.sample_observations

TMI.sample_observations

TMI.section

TMI.steady_inversion

TMI.surfaceindex

TMI.surfaceorigin

TMI.surfacepatch

TMI.tracerinit

TMI.trackpathways

TMI.updatelinearindex

TMI.vec2fld

TMI.volumefilled

TMI.watermassmatrix

TMI.watermassmatrixXYZ

TMI.watermassmatrixZYX

TMI.wetlocation

TMI.cartesianindexXYZ — Method

function cartesianindexXYZ(wet)
Read and assemble the grid coordinates
according to a 3D tracer in x,y,z order

Arguments

wet: BitArray logical mask for wet points

Output

I: 3D Cartesian indices

TMI.cartesianindexZYX — Method

function cartesianindexZYX(file)
Read and assemble the grid coordinates
according to the legacy MATLAB code (z,y,x order).

Arguments

file: TMI NetCDF file name

Output

grid: TMI grid coordinates

TMI.config — Method

function config(url,inputdir)

Arguments

TMIversion: TMI version for water-mass/circulation model

Output

A: TMI steady-state water-mass matrix

Alu: LU decomposition of A

γ: TMI grid properties

TMIfile: TMI file name

TMI.control2state! — Method

function control2state!(c,u,γ)
Add surface control vector to existing 3D field

Arguments

c:: state field, 3d tracer field with NaN on dry points, modified by function

u:: surface control vector

wet::BitArray mask of ocean points

TMI.control2state! — Method

function control2state!(c,u,γ)
Add surface control vector to existing 3D field

Arguments

c:: state field, 3d tracer field with NaN on dry points, modified by function

u:: surface control vector

wet::BitArray mask of ocean points

TMI.control2state — Method

function control2state(tracer2D,γ)
turn 2D surface field into 3D field with zeroes below surface

Arguments

tracer2D:: 2D surface tracer field

wet::BitArray mask of ocean points

Output

tracer3D:: 3d tracer field with NaN on dry points

TMI.control2state — Method

function control2state(u,γ)
turn surface control vector into 3D field with zeroes below surface

Arguments

u:: surface control vector

wet::BitArray mask of ocean points

Output

tracer3D:: 3d tracer field with NaN on dry points

TMI.costfunction! — Method

function costfunction!(J,gJ,u,Alu,dfld,yfld,Wⁱ,wis,Q⁻,γ)
squared model-data misfit for pointwise data
controls are a vector input for Optim.jl
Issue: couldn't figure out how to nest with costfunction_obs!

Arguments

J: cost function of sum of squared misfits

gJ: derivative of cost function wrt to controls

u: controls, vector format

Alu: LU decomposition of water-mass matrix

dfld: model constraints

y: pointwise observations

Wⁱ: inverse of W weighting matrix for observations

wis: weights for interpolation (data sampling, E)

locs: data locations (lon,lat,depth)

Q⁻: weights for control vector

γ: grid

TMI.costfunction — Method

function costfunction(J,gJ,u,Alu,dfld,yfld,Wⁱ,wis,Q⁻,γ)
squared model-data misfit for pointwise data
controls are a vector input for Optim.jl
Issue: couldn't figure out how to nest with costfunction_obs!

Arguments

u: controls, vector format

Alu: LU decomposition of water-mass matrix

dfld: model constraints

y: pointwise observations

Wⁱ: inverse of W weighting matrix for observations

wis: weights for interpolation (data sampling, E)

locs: data locations (lon,lat,depth)

Q⁻: weights for control vector

γ: grid

Output

J: cost function of sum of squared misfits

gJ: derivative of cost function wrt to controls

TMI.costfunction_obs! — Method

function costfunction_obs(u,Alu,y,d,Wⁱ,wet)
squared model-data misfit for gridded data
controls are a vector input for Optim.jl

Arguments

J: cost function of sum of squared misfits

gJ: derivative of cost function wrt to controls

u: controls, vector format

Alu: LU decomposition of water-mass matrix

d: model constraints

y: observations on grid

Wⁱ: inverse of W weighting matrix for observations

γ: grid

TMI.costfunction_obs! — Method

function costfunction_obs!(J,gJ,u,Alu,dfld,yfld,Wⁱ,wis,locs,γ)
squared model-data misfit for pointwise data
controls are a vector input for Optim.jl

Arguments

J: cost function of sum of squared misfits

gJ: derivative of cost function wrt to controls

u: controls, vector format

Alu: LU decomposition of water-mass matrix

dfld: model constraints

y: pointwise observations

Wⁱ: inverse of W weighting matrix for observations

wis: weights for interpolation (data sampling, E)

locs: data locations (lon,lat,depth)

γ: grid

TMI.costfunction_obs — Method

function costfunction_obs(u,Alu,dfld,yfld,Wⁱ,γ)
squared model-data misfit for gridded data
controls are a vector input for Optim.jl

Arguments

u: controls, vector format

Alu: LU decomposition of water-mass matrix

y: observations on grid

d: model constraints

Wⁱ: inverse of W weighting matrix for observations

wet: BitArray ocean mask

Output

J: cost function of sum of squared misfits

gJ: derivative of cost function wrt to controls

TMI.costfunction_obs — Method

function costfunction_obs(u,Alu,dfld,yfld,Wⁱ,wis,locs,γ)
squared model-data misfit for pointwise data
controls are a vector input for Optim.jl

Arguments

u: controls, vector format

Alu: LU decomposition of water-mass matrix

y: pointwise observations

d: model constraints

Wⁱ: inverse of W weighting matrix for observations

wis: weights for interpolation

locs: data locations

γ: grid

Output

J: cost function of sum of squared misfits

gJ: derivative of cost function wrt to controls

TMI.depthindex — Method

function depthindex(I) 
Get the k-index (depth level) from the Cartesian index

TMI.download — Method

function download(url,inputdir)
Read and assemble all TMI inputs.

Arguments

url: Google Drive location of TMI input

inputdir: input directory location to store file

Output

none

TMI.dyeplot — Method

function dyeplot
Plot of dye in ocean

Arguments

lat: latitude arrays

depth: depth array

vals: lat x depth value array

lims: contour levels

TMI.filterdata — Method

function filterdata(u₀,Alu,y,d₀,W⁻,γ) Find the distribution of a tracer given: (a) the pathways described by A or its LU decomposition Alu, (b) first-guess boundary conditions and interior sources given by d₀, (c) perturbations to the surface boundary condition u₀ that best fits observations, y, according to the cost function, J = (ỹ - y)ᵀ W⁻¹ (ỹ - y) subject to Aỹ = d₀ + Γ u₀. W⁻ is a (sparse) weighting matrix. See Supplementary Section 2, Gebbie & Huybers 2011.

Arguments

u₀:

Alu:

d₀: first guess of boundary conditions and interior sources

y: observations on 3D grid

W⁻: weighting matrix best chosen as inverse error covariance matrix

fg!: compute cost function and gradient in place

γ: grid

TMI.fld2vec — Method

function fld2vec
Transfer 3D field with accounting for ocean bathymetry to a vector without land points

Arguments

field: field in 3d form including land points (NaN)

I: cartesian indices of ocean points

Output

vector: field in vector form (no land points)

TMI.gdriveurl — Method

function gdriveurl(TMIversion)
placeholder function to give location (URL) of Google Drive input
in the future, consider a struct or Dict that describes all TMI versions.

Arguments

TMIversion: version of TMI water-mass/circulation model

Output

url: location (URL) for download

TMI.gridprops — Method

function gridprops(file)
Read and assemble the grid properties.

Arguments

file: TMI NetCDF file name

Output

grid: TMI grid coordinates

TMI.horizontaldistance — Method

function horizontaldistance(loc,γ)
return the Cartesian index and linear index 
of the nearest N neighbors

Arguments

loc: 3-tuple of lon,lat,depth location

γ: TMI.grid

Output

hordist: horizontal distance to nearest tracer grid points

TMI.interpindex — Method

function interpindex(loc,γ) Weights for linear interpolation. The derivative of linear interpolation is needed in sensitivity studies. ReverseDiff.jl could find this quantity automatically. Instead we dig into the Interpolations.jl package to find the weights that are effectively the partial derivatives of the function.

Arguments

c: a temporary tracer field, would be nice to make it unnecessary

loc: (lon,lat,depth) tuple of a location of interest

γ: TMI grid

Output

δ: weights on a 3D tracer field grid

TMI.interpweights — Method

function interpweights(loc,γ) Weights for linear interpolation. The derivative of linear interpolation is needed in sensitivity studies. ReverseDiff.jl could find this quantity automatically. Instead we dig into the Interpolations.jl package to find the weights that are effectively the partial derivatives of the function.

Arguments

loc: (lon,lat,depth) tuple of a location of interest

γ: TMI grid

Output

δ: weights on a 3D tracer field grid

TMI.linearindexXYZ — Method

function linearindexXYZ(file)
Read and assemble the grid coordinates.

Arguments

wet: 3D mask for wet points

Output

R: array of linear indices, but not a LinearIndices type

TMI.nearestneighbor — Function

function nearestneighbor(loc,γ)
return the Cartesian index and linear index 
of the nearest N neighbors

Arguments

loc: 3-tuple of lon,lat,depth location

γ: TMI.grid

Output

Inn: Cartesian indices of nearest neighbor

#- Rnn: linear indices of nearest neighbor, Removed from code

TMI.nearestneighbormask — Function

function nearestneighbormask
Make a 3D tracer field that is 1 at location 
of nearest neighbor, 0 elsewhere

Arguments

loc: location in a 3-tuple (lon,lat,depth)

γ: TMI.grid

Output

δ: nearest neighbor mask 3D field

TMI.plotextent — Method

function plotextent
Generate image showing user-specified ROI

Arguments

latbox: in format [latstart, latstop]

lonbox: in format [lonstart, lonstop]

TMI.readtracer — Method

function readtracer(file,tracername)
Read and assemble the water-mass matrix.

Arguments

file: TMI NetCDF file name

tracername: name of tracer

Output

c: 3D tracer field

TMI.regeneratedphosphate — Method

function regeneratedphosphate(TMIversion)
Regenerated (i.e., accumulated, remineralized) phosphate

Arguments

TMIversion: version of TMI water-mass/circulation model

Output

PO₄ᴿ: regenerated phosphate

γ: TMI grid

TMI.sample_observations — Method

function sample_observations(TMIversion,variable,locs)
Synthetic observations that are a contaminated version of real observations
This version: observations with random (uniform) spatial sampling

Arguments

TMIversion::String: version of TMI water-mass/circulation model

variable::String: variable name to use as template

N: number of observations

Output

y: contaminated observations on 3D grid

W⁻: appropriate weighting (inverse covariance) matrix for these observations,

ytrue: uncontaminated observations, 3D field

locs: 3-tuples of locations for observations

wis: weighted indices for interpolation to locs sites

TMI.sample_observations — Method

function sample_observations(TMIversion,variable)
Synthetic observations that are a contaminated version of real observations
This version: gridded observations

Arguments

TMIversion::String: version of TMI water-mass/circulation model

variable::String: variable name to use as template

Output

y: contaminated observations on 3D grid

W⁻: appropriate weighting (inverse covariance) matrix for these observations,

θtrue: real observations, 3D field

TMI.section — Method

function section
View latitude-depth slice of field

Arguments

c: 3d tracer field

lon: longitude of section

γ: TMI.grid

Output

csection: 2d slice of field

TMI.steady_inversion — Method

function steady_inversion(u,Alu,d,γ.wet)
invert for a steady-state tracer distribution

Arguments

u: controls, vector format

Alu: LU decomposition of water-mass matrix

d: model constraints

wet: BitArray ocean mask

Output

c: steady-state tracer distribution

TMI.surfaceindex — Method

function surfaceindex(I) 
Get the vector-index where depth level == 1 and it is ocean.

TMI.surfaceorigin — Method

function surfaceorigin(TMIversion,loc)
 Find the surface origin of water for some interior box 
 This is equivalent to solving a sensitivity problem:
 The mass fraction at a location `loc` of interest is 
`c[loc] = δᵀ c`, where `δ` samples the location of the global mass-fraction variable, c.
Then the sensitivity of `c[loc]` is: d(c[loc])/d(d) = A⁻ᵀ δ.
The derivative is solved using the constraint: Ac = d.
The sensitivity is exactly the mass fraction originating from each source.      
This problem is mathematically similar to determining how the ocean is filled.

Arguments

TMIversion: version of TMI water-mass/circulation model

loc: location (lon,lat,depth) of location of interest

Output

origin: surface map of fraction of source water for a given location

γ: TMI grid

TMI.surfacepatch — Method

function surfacepatch
Make a surface boundary condition
with a rectangular patch

Arguments

lonbox: longitudes of box edges

latbox: latitudes of box edges

γ: TMI.grid

Output

d: vector that describes surface patch

TMI.tracerinit — Function

function tracerinit(wet,ltype=Float64)
  initialize tracer field on TMI grid
perhaps better to have a tracer struct and constructor

Arguments

wet::BitArray mask of ocean points

Output

d:: 3d tracer field with NaN on dry points

TMI.trackpathways — Method

function trackpathways(TMIversion,latbox,lonbox)
Track the pathways of a user-defined water mass.
 Steps: (a) define the water mass by a rectangular surface patch dyed with passive tracer concentration of         (b) propagate the dye with the matrix A, with the result being the fraction of water originating from the surface region.
 See Section 2b of Gebbie & Huybers 2010, esp. eqs. (15)-(17).

Arguments

TMIversion: version of TMI water-mass/circulation model

latbox: min and max latitude of box

lonbox: min and max longitude of box

Output

c: fraction of water from surface source

γ: TMI grid

TMI.updatelinearindex — Method

function updatelinearindex(izyx,Izyx,R)
Linear index translated from z,y,x to x,y,z accounting

Arguments

izyx: index of interest in z,y,x accounting

Izyx: wet Cartesian Index for z,y,x

R: Linear indices for x,y,z

Output

ixyz: index of interest in x,y,z accounting

TMI.vec2fld — Method

function vec2fld
Transfer a vector to a 3D field with accounting for ocean bathymetry

Arguments

vector: field in vector form (no land points)

I: cartesian indices of ocean points

Output

field: field in 3d form including land points (NaN)

TMI.volumefilled — Method

function volumefilled(TMIversion)
Find the ocean volume that has originated from each surface box.
 This is equivalent to solving a sensitivity problem:
 The total volume is V = vᵀ c , where v is the volume of each box 
 and c is the fraction of volume from a given source which
 satisfies the equation A c = d.                     
 Next, dV/d(d) = A⁻ᵀ v, and dV/d(d) is exactly the volume originating from each source.

 See Section 3 and Supplementary Section 4, Gebbie & Huybers 2011.

Arguments

TMIversion: version of TMI water-mass/circulation model

Output

volume: global ocean volume filled by a surface region

TMI.watermassmatrix — Method

function watermassmatrix(file)
Read and assemble the water-mass matrix.

Arguments

file: TMI NetCDF file name

Output

A: water-mass matrix

TMI.watermassmatrixXYZ — Method

    function watermassmatrixXYZ(file,R)
Read and assemble the water-mass matrix from MATLAB.
Transfer to updated x,y,z version

Arguments

file: TMI NetCDF file name

γ: TMI grid

Output

A: water-mass matrix

TMI.watermassmatrixZYX — Method

function watermassmatrixZYX(file)
Read and assemble the water-mass matrix.
Legacy version from MATLAB.

Arguments

file: TMI NetCDF file name

Output

A: water-mass matrix

TMI.wetlocation — Method

function wetlocation(γ)
Get (lon,lat,depth) tuples of wet locations.
Allow a location to be wet if at least one out of 8 nearby gridpoints is wet.
Certainly "wet" gridpoints could be defined more strictly.

Arguments

γ: TMI.grid

Output

loc: lon,lat,depth tuple