Configuration

Reading input files and configuration for TMI.

TMI.config_from_nc — Function

function config_from_nc(TMIversion)
Configure TMI environment from NetCDF input file.

Arguments

TMIversion: TMI version for water-mass/circulation model

Output

A: TMI steady-state water-mass matrix

Alu: LU decomposition of A

γ: TMI grid properties

TMIfile: TMI file name

TMI.config_from_mat — Function

Configure TMI environment from original MATLAB output

TMI.watermassmatrix — Function

function watermassmatrix(file)
Read and assemble the water-mass matrix.

Arguments

file: TMI NetCDF or MATLAB file name

Output

A: water-mass matrix

TMI.matrix_zyx2xyz — Function

    function matrix_zyx2xyz(TMIfile,Azyx,γ)

Transfer zyx format water-mass matrix A to xyz format

Arguments

Azyx: water-mass matrix in zyx format

γ: TMI grid

Output

Axyz: water-mass matrix in xyz format

TMI.circulationmatrix — Function

function circulationmatrix(file,γ)
Read and assemble the circulation matrix from MATLAB.
Transfer to updated x,y,z version

Arguments

file: TMI MATLAB file name

γ: TMI grid

Output

L: circulation matrix in xyz format

function circulationmatrix(file,A,γ)
Read and assemble the circulation matrix from the efficient storage of A and F₀ variables.

Arguments

file: TMI MATLAB file name

A: TMI water-mass matrix

γ: TMI grid

Output

L: circulation matrix in xyz format

TMI.boundarymatrix — Function

    function boundarymatrix(file,γ)
Read and assemble the boundary matrix from MATLAB.
Transfer to updated x,y,z version

Arguments

file: TMI MATLAB file name

γ: TMI grid

Output

B: boundary condition matrix

TMI.updatelinearindex — Function

function updatelinearindex(izyx,Izyx,R)
Linear index translated from z,y,x to x,y,z accounting

get Izyx Cartesian index stored from legacy MATLAB code

Arguments

izyx: index of interest in z,y,x accounting

Izyx: wet Cartesian Index for z,y,x

R: Linear indices for x,y,z

Output

ixyz: index of interest in x,y,z accounting

TMI.surfaceregion — Function

 function surfaceregion(TMIversion::String,region::String,γ::Grid)::BoundaryCondition

Read an oceanographically-relevant surface region from NetCDF file. (Also could be read from mat file.)
Return a BoundaryCondition

TMI.ncurl — Function

function ncurl(TMIversion)
placeholder function to give location (URL) of NetCDF Google Drive input
in the future, consider a struct or Dict that describes all TMI versions.

Arguments

TMIversion: version of TMI water-mass/circulation model

Output

url: location (URL) for download

TMI.maturl — Function

function maturl(TMIversion)
Find *mat file here.
placeholder function to give location (URL) of Google Drive input
in the future, consider a struct or Dict that describes all TMI versions.

Arguments

TMIversion: version of TMI water-mass/circulation model

Output

url: location (URL) for download

TMI.config2nc — Function

Save TMI configuration to NetCDF format for non-proprietary access

TMI.griddicts — Function

Save grid dictionaries of attributes for writing to NetCDF file

TMI.matfields2nc — Function

Read 3D fields from mat file and save to NetCDF file.

TMI.fieldsatts — Function

All variable names and attributes. Useful for writing NetCDF files.

TMI.regions2nc — Function

Read vectors from mat file, translate to 3D, and save surface field to NetCDF file.

Missing docstring.

Missing docstring for TMI.watermassmatrix2nc. Check Documenter's build log for details.

TMI.optim2nc — Function

Save optimization parameters to NetCDF file)

Future considerations: split into 2 functions

read from mat

save to nc

TMI.circulationmatrix2nc — Function

Save circulation matrix L to NetCDF file.

TMI.boundarymatrix2nc — Function

Save boundary matrix for transient model to NetCDF file

TMI.grid2nc — Function

Put grid properties (Cartesian index) into NetCDF file

TMI.cartesianindex — Function

function cartesianindex(file)
Read and assemble the grid coordinates
according to the legacy MATLAB code (z,y,x order).

Arguments

file: TMI NetCDF file name

Output

I: TMI Cartesian index for wet points

function cartesianindex(wet)
Read and assemble the grid coordinates
according to a 3D tracer in x,y,z order

Arguments

wet: BitArray logical mask for wet points

Output

I: 3D Cartesian indices

TMI.gridprops — Function

function gridprops(file)
Read and assemble the grid properties.

Arguments

file: TMI NetCDF file name

Output

grid: TMI grid coordinates