Grid

Handling the grid for TMI.

TMI.Grid — Type

struct Grid

TMI grid with accounting for wet/dry points

TMI.Field — Type

struct Field

This structure permits the grid to be 
automatically passed to functions with
the tracer field.

This structure assumes the Tracer type to be 
three-dimensional.

Missing docstring.

Missing docstring for TMI.cartesianindex. Check Documenter's build log for details.

TMI.linearindex — Function

function linearindex(wet)
Read and assemble the grid coordinates.

Arguments

wet: 3D mask for wet points

Output

R: array of linear indices, but not a LinearIndices type

Missing docstring.

Missing docstring for TMI.gridprops. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TMI.matrix_zyx2xyz. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TMI.updatelinearindex. Check Documenter's build log for details.

TMI.cellarea — Function

Horizontal area of grid cell

TMI.cellvolume — Function

Volume of each grid cell.

Missing docstring.

Missing docstring for TMI.layerthickness. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TMI.zonalgriddist. Check Documenter's build log for details.

TMI.surfacepatch — Function

function surfacepatch
Make a surface boundary condition
with a rectangular patch

Arguments

lonbox: longitudes of box edges

latbox: latitudes of box edges

γ: TMI.grid

Output

d: vector that describes surface patch

TMI.depthindex — Function

function depthindex(I) 
Get the k-index (depth level) from the Cartesian index

TMI.lonindex — Function

function lonindex(I) 
Get the i-index (lon index) from the Cartesian index

TMI.latindex — Function

function latindex(I) 
Get the j-index (latitude index) from the Cartesian index

Missing docstring.

Missing docstring for TMI.findindex. Check Documenter's build log for details.

TMI.surfaceindex — Function

function surfaceindex(I) 
Get the vector-index where depth level == 1 and it is ocean.

TMI.northindex — Function

function northindex(I) 
Get the vector index on the northern open boundary

TMI.eastindex — Function

function eastindex(I) 
Get the vector index on the northern open boundary

TMI.southindex — Function

function southindex(I) 
Get the vector-index on the southern open boundary

TMI.westindex — Function

function westindex(I) 
Get the vector index on the western open boundary

TMI.horizontaldistance — Function

function horizontaldistance(loc,γ)
return the Cartesian index and linear index 
of the nearest N neighbors

Arguments

loc: 3-tuple of lon,lat,depth location

γ: TMI.grid

Output

hordist: horizontal distance to nearest tracer grid points

TMI.interpindex — Function

function interpindex(loc,γ) Weights for linear interpolation. The derivative of linear interpolation is needed in sensitivity studies. ReverseDiff.jl could find this quantity automatically. Instead we dig into the Interpolations.jl package to find the weights that are effectively the partial derivatives of the function.

Arguments

c: a temporary tracer field, would be nice to make it unnecessary

loc: (lon,lat,depth) tuple of a location of interest

γ: TMI grid

Output

δ: weights on a 3D tracer field grid

TMI.shiftloc — Function

function shiftloc(loc)

sometimes loc longitudes are outside of grid due to different conventions
assumption: 360° shift is enough to get back on grid

TMI.interpweights — Function

function interpweights(loc,γ) Weights for linear interpolation. The derivative of linear interpolation is needed in sensitivity studies. ReverseDiff.jl could find this quantity automatically. Instead we dig into the Interpolations.jl package to find the weights that are effectively the partial derivatives of the function.

Arguments

loc: (lon,lat,depth) tuple of a location of interest

γ: TMI grid

Output

δ: weights on a 3D tracer field grid